requirements.txt
aiogram>=3.0
openai>=1.0
python-dotenv
